requirements.txt
Flask
gunicorn
Werkzeug
numpy
opencv-python
ultralytics
easyocr
torch
symspellpy
google-generativeai
dotenv